fix: 락 관리 이벤트 consumer 적용 및 db 스키마 수정
