+++ app/domain/common-domain/build.gradle
@@ -1,6 +1,4 @@
 dependencies {
     //ulid
     implementation 'com.github.f4b6a3:ulid-creator:5.2.3'
-    //retry
-    implementation 'org.springframework.retry:spring-retry'
 }
