+++ app/domain/show-domain/src/main/java/org/example/dto/viewcount/ShowViewCountEvent.java
@@ -3,4 +3,8 @@ package org.example.dto.viewcount;
 import java.util.UUID;
 
 public record ShowViewCountEvent(UUID showId) {
+
+    public static ShowViewCountEvent from(UUID showId) {
+        return new ShowViewCountEvent(showId);
+    }
 }
